refactor: further modularized code base
